\name{stmCorrViz-package}
\alias{stmCorrViz-package}
\docType{package}
\title{
Hierarchical Correlation View of STM Models
}
\description{
A tool for hierarchical visualizations of the Structural Topic Model (STM) of Roberts, Stewart, and Tingley. The package allows the user to generate interactive visualizations highlighting correlations among topics, and can be used to view the model at differing levels of complexity.
}
\details{
\tabular{ll}{
Package: \tab stmCorrViz\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-02-14\cr
License: \tab GPL (>= 2) \cr
}
}
\author{
Antonio Coppola, Margaret E. Roberts, Brandon M. Stewart, and Dustin Tingley

Maintainer: Antonio Coppola <acoppola@college.harvard.edu>
}
\references{
Margaret E. Roberts, Brandon M. Stewart and Dustin Tingley (2014). \href{http://structuraltopicmodel.com/}{stm: R Package for Sturctural Topic Models}.
}

