% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmCorrViz_htmlwidget.R
\name{stmCorrViz_htmlwidget-shiny}
\alias{renderStmCorrViz_htmlwidget}
\alias{stmCorrViz_htmlwidget-shiny}
\alias{stmCorrViz_htmlwidgetOutput}
\title{Shiny bindings for stmCorrViz_htmlwidget}
\usage{
stmCorrViz_htmlwidgetOutput(outputId, width = "100\%", height = "400px")

renderStmCorrViz_htmlwidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a stmCorrViz_htmlwidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using stmCorrViz_htmlwidget within Shiny
applications and interactive Rmd documents.
}

