\name{stmCorrViz-package}
\alias{stmCorrViz-package}
\docType{package}
\title{
Hierarchical Correlation View of STM Models
}
\description{
Generates an interactive visualization of topic correlations/hierarchy in a Structural Topic Model (STM) of Roberts, Stewart, and Tingley.  The package performs a hierarchical clustering of topics which are then exported to a JSON object and visualized using D3.
}
\details{
\tabular{ll}{
Package: \tab stmCorrViz\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-02-14\cr
License: \tab GPL (>= 2) \cr
}
}
\author{
Antonio Coppola, Margaret E. Roberts, Brandon M. Stewart, and Dustin Tingley

Maintainer: Antonio Coppola <acoppola@college.harvard.edu>
}
\references{
Margaret E. Roberts, Brandon M. Stewart and Dustin Tingley (2014). \href{http://structuraltopicmodel.com/}{stm: R Package for Sturctural Topic Models}.
}

